meeting_prep_gamma_github/requirements.txt
# Core
streamlit>=1.40.0
python-dotenv>=1.0.1
pydantic>=2.9.0

# OpenAI Agents SDK
openai>=1.0.0,<2.0.0
openai-agents>=0.0.1

# Document parsing
pdfplumber>=0.11.4
pymupdf4llm>=0.0.1
python-docx>=1.1.2
openpyxl>=3.1.5
pandas>=2.2.0

# HTTP / APIs
requests>=2.32.0
httpx>=0.27.0

# Utilities
python-multipart>=0.0.9
